+++ packages/nativescript/src/generators/application/generator.ts
@@ -6,63 +6,26 @@ import {
     addProjectConfiguration,
     formatFiles,
     GeneratorCallback,
-    getWorkspaceLayout,
-    names,
     NxJsonProjectConfiguration,
-    offsetFromRoot,
     ProjectConfiguration,
     Tree
 } from '@nrwl/devkit';
 
 /**
- * 3rd-party imports
- */
-import * as path from 'path';
-
-/**
  * Internal imports
  */
 import { NativeScriptAppGeneratorSchema } from './schema';
-import { generateFiles } from '../generate-files';
 import {
     DEPENDENCIES,
     DEV_DEPENDENCIES
 } from './dependencies';
+import { NormalizedSchema } from 'packages/schemas';
+import {
+    addFiles,
+    normalizeOptions
+} from 'packages/common.utils';
 
-/**
- * TypeScript entities and constants
- */
-interface NormalizedSchema extends NativeScriptAppGeneratorSchema {
-    projectName: string;
-    projectRoot: string;
-    projectDirectory: string;
-    parsedTags: string[];
-}
-
-function normalizeOptions(
-    host: Tree,
-    options: NativeScriptAppGeneratorSchema
-): NormalizedSchema {
-    const name: string = names(options.name).fileName;
-    const projectDirectory: string = options.directory
-        ? `${names(options.directory).fileName}/${name}`
-        : name;
-    const projectName: string = projectDirectory.replace(new RegExp('/', 'g'), '-');
-    const projectRoot = `${getWorkspaceLayout(host).appsDir}/${projectDirectory}`;
-    const parsedTags: string[] = options.tags
-        ? options.tags.split(',').map((s) => s.trim())
-        : [];
-
-    return {
-        ...options,
-        projectName,
-        projectRoot,
-        projectDirectory,
-        parsedTags
-    };
-}
-
-function generateProjectConfiguration(normalizedOptions: NormalizedSchema): ProjectConfiguration & NxJsonProjectConfiguration {
+function generateProjectConfiguration(normalizedOptions: NormalizedSchema<NativeScriptAppGeneratorSchema>): ProjectConfiguration & NxJsonProjectConfiguration {
     return {
         root: normalizedOptions.projectRoot,
         projectType: 'application',
@@ -112,28 +75,11 @@ function generateProjectConfiguration(normalizedOptions: NormalizedSchema): Proj
     };
 }
 
-function addFiles(host: Tree, templatesFolder: string, options: NormalizedSchema): void {
-    const templateOptions: Record<string, string | string[]> = {
-        ...options,
-        ...names(options.name),
-        offsetFromRoot: offsetFromRoot(options.projectRoot),
-        tmpl: '',
-        dot: '.'
-    };
-
-    generateFiles(
-        host,
-        path.join(__dirname, 'templates', templatesFolder),
-        options.projectRoot,
-        templateOptions
-    );
-}
-
 export default async function (
     host: Tree,
     options: NativeScriptAppGeneratorSchema
 ): Promise<GeneratorCallback> {
-    const normalizedOptions: NormalizedSchema = normalizeOptions(host, options);
+    const normalizedOptions: NormalizedSchema<NativeScriptAppGeneratorSchema> = normalizeOptions(host, options);
     addProjectConfiguration(host, normalizedOptions.projectName, generateProjectConfiguration(normalizedOptions));
 
     /*
@@ -147,10 +93,10 @@ export default async function (
     );
 
     // Generating files from templates according to the app's type
-    addFiles(host, `templates-${options.type}`, normalizedOptions);
+    addFiles(host, normalizedOptions, `templates-${options.type}`);
 
     // Generating App_Resources files from templates
-    addFiles(host, 'templates-common', normalizedOptions);
+    addFiles(host, normalizedOptions, 'templates-common',);
 
     // Formatting files according to Prettier
     await formatFiles(host);
