+++ packages/nativescript/src/generators/application/schema.json
@@ -5,6 +5,11 @@
   "title": "",
   "type": "object",
   "properties": {
+    "projectType": {
+      "type": "string",
+      "description": "The type of project this generator will build, either 'application' or 'library'",
+      "default": "application"
+    },
     "name": {
       "type": "string",
       "description": "",
