+++ packages/nest-sls/src/generators/application/generator.ts
@@ -3,75 +3,28 @@
  */
 import {
     formatFiles,
-    generateFiles,
-    getWorkspaceLayout,
-    names,
-    offsetFromRoot,
     Tree
 } from '@nrwl/devkit';
 import { applicationGenerator } from '@nrwl/nest';
-import * as path from 'path';
 
 /**
  * Internal imports
  */
-import { NestSlsApplicationSchema } from './schema';
+import { NestSlsAppGeneratorSchema } from './schema';
+import {
+    addFiles,
+    normalizeOptions
+} from 'packages/common.utils';
 
-/**
- * TypeScript entities and constants
- */
-interface NormalizedSchema extends NestSlsApplicationSchema {
-    projectName: string;
-    projectRoot: string;
-    projectDirectory: string;
-    parsedTags: string[];
-}
-
-function normalizeOptions(
-    host: Tree,
-    options: NestSlsApplicationSchema
-): NormalizedSchema {
-    const name = names(options.name).fileName;
-    const projectDirectory = options.directory
-        ? `${names(options.directory).fileName}/${name}`
-        : name;
-    const projectName = projectDirectory.replace(new RegExp('/', 'g'), '-');
-    const projectRoot = `${getWorkspaceLayout(host).appsDir}/${projectDirectory}`;
-    const parsedTags = options.tags
-        ? options.tags.split(',').map((s) => s.trim())
-        : [];
-
-    return {
-        ...options,
-        projectName,
-        projectRoot,
-        projectDirectory,
-        parsedTags,
-    };
-}
-
-function addFiles(host: Tree, options: NormalizedSchema) {
-    const templateOptions = {
-        ...options,
-        ...names(options.name),
-        offsetFromRoot: offsetFromRoot(options.projectRoot),
-        template: '',
-    };
-    generateFiles(
-        host,
-        path.join(__dirname, 'files'),
-        options.projectRoot,
-        templateOptions
-    );
-}
-
-export default async function (host: Tree, options: NestSlsApplicationSchema) {
+export default async function (host: Tree, options: NestSlsAppGeneratorSchema) {
     const normalizedOptions = normalizeOptions(host, options);
 
-    // TODO: Use the app generator from the @nrwl/nest plugin and add necessary files
-    // https://nx.dev/latest/angular/core-concepts/nx-devkit#composing-generators
+    // Generating a basic Nest app
     await applicationGenerator(host, normalizedOptions);
 
+    // Generating files from templates
     addFiles(host, normalizedOptions);
+
+    // Formatting files according to Prettier
     await formatFiles(host);
 }
